changed isEqual method
What changes you have made?
.....
it is now checking equality of two object's filds and also the
constructore from which they are made.

Why this change is necessary?
....
this change was necessary for batter testing of two objects.

What is the impact of the change?
.....
this is helpful for testing two object's equality.